// src/state/slices/opportunitiesSlice/types.ts
export enum DefiProvider {
  Idle = 'Idle',
  ShapeShift = 'ShapeShift',
  EthJinxStaking = 'ETH/JINX Staking',
  UniV2 = 'Uniswap V2',
  CosmosSdk = 'Cosmos SDK',
  OsmosisLp = 'Osmosis LP',
  ThorchainSavers = 'THORChain Savers',
}

export type DefiProviderMetadata = {
  provider: string
  icon: string
  color: string
  url?: string
}
// ...
export type StakingIdBase = Nominal<string, 'StakingId'> & AssetId
// Sometimes, an AssetIdish (i.e a a CAIP-19) is not enough to uniquely identify a staking opportunity
// So we add a StakingKey to the StakingId to make it unique
export type StakingKey = Nominal<string, 'StakingKey'>
export type StakingId = StakingIdBase | `${StakingIdBase}#${StakingKey}`
// The AccountId of the LP contract in the form of chainId:accountAddress
export type LpId = Nominal<string, 'LpId'> & AssetId